Tan cong/py-script/pyproject.toml
[tool.poetry]
name = "pycharmprojects"
version = "0.1.0"
description = ""
authors = ["truong nguyen <truongjenny46@mgmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
